requirements.txt
yfinance==0.2.14
matplotlib==3.7.1
tk==0.1.0
pandastable==0.13.1
numpy>=1.24.0